% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_animated_letters.R
\name{add_animated_letters}
\alias{add_animated_letters}
\title{Apply css animation}
\usage{
add_animated_letters(
  text = "You can write here a sentence.",
  animation = "neon",
  color = "#FFF",
  size = "1em"
)
}
\arguments{
\item{text}{Character Text to which the animation should be applied letter by letter.}

\item{animation}{Character. The kind of animation: "bouncing", "infire", or "neon".}

\item{color}{Character. CSS specification of the font size.}

\item{size}{Character. CSS specification of the font color.}
}
\value{
Write the quarto code for the animation.
}
\description{
Apply css animation separately for each letter of the entered text.
}
\author{
Nicolas Mangin
}
