% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meme.R
\name{add_meme}
\alias{add_meme}
\title{Insert a slide}
\usage{
add_meme(
  text_above = "",
  image = "",
  text_below = "",
  font_size = 2.5,
  transback = "slide",
  transdata = "fade"
)
}
\arguments{
\item{text_above}{Character. Sentence written above the image or video.}

\item{image}{Character. Address of the background image.}

\item{text_below}{Character. Sentence written below the image or video.}

\item{font_size}{Numeric. Size of the font in "em".}

\item{transback}{Character. Type of transition for the background: fade, slide, convex, concave, or zoom.}

\item{transdata}{Character. Type of transition for data: fade, slide, convex, concave, or zoom.}
}
\value{
Character. Write the first two rows of a RevealJS slide formatted for rmarkdown.
}
\description{
Function writing a revealjs slide according to the user specifications.
}
\author{
Nicolas Mangin
}
