% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rating.R
\name{add_rating}
\alias{add_rating}
\title{Add star rating}
\usage{
add_rating(
  sourcename,
  filename,
  message =
    "Please let us know below how clear and understandable this page was for you:  "
)
}
\arguments{
\item{sourcename}{Character. Name of the publication (blogpost, textbook, or presentation)}

\item{filename}{Character. Code and language of the document.}

\item{message}{Character. Comment entered by the reader.}
}
\value{
Write the HTML code for the star rating in the page
}
\description{
Function writing the HTML code to include the page star rating. Note that this requires to adapt the javascript function so that the comment goes to the appropriate Google form.
}
\author{
Nicolas Mangin
}
