% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_video.R
\name{add_video}
\alias{add_video}
\title{Embed video}
\usage{
add_video(vid, host = "Youtube")
}
\arguments{
\item{vid}{Character. ID of the video.}

\item{host}{Character. Name of the hosting platform. For now, only "Youtube" is supported.}
}
\value{
Write the HTML code embedding the videos.
}
\description{
Function embedding a video as an ifram in an html page
}
\author{
Nicolas Mangin
}
