% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bs3_dialog.R
\name{bs3_dialog}
\alias{bs3_dialog}
\title{Custom dialog box function for \code{\link[editR:editR-package]{editR}}}
\usage{
bs3_dialog(id, title, trigger, ...)
}
\arguments{
\item{id}{A unique identifier for the dialog box.}

\item{title}{The title to appear at the top of the dialog box.}

\item{trigger}{The id of a button or link that will open the dialog box.}

\item{...}{UI elements to include within the dialog box (e.g. \code{\link[shiny:actionButton]{actionButtons}}).}
}
\value{
This function return the html code for a dialog box.
}
\description{
This function creates custom Twitter Bootstrap 3 dialog boxes for \code{\link[editR:editR-package]{editR}}.
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

