% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bs3_dropdown.R
\name{bs3_dropdown}
\alias{bs3_dropdown}
\title{Custom dropdown menu function for \code{\link[editR:editR-package]{editR}}}
\usage{
bs3_dropdown(name = "My menu", navs)
}
\arguments{
\item{name}{The name of the dropdown menu.}

\item{navs}{A list of UI elements to include in the dropdown menu (e.g. \code{\link[shiny:actionLink]{actionLinks}}).}
}
\value{
This function return the html code for a Twitter Bootstrap 3 dropdown menu.
}
\description{
This function creates custom Twitter Bootstrap 3 dropdown menus for \code{\link[editR:editR-package]{editR}}.
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

