% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bs3_navbar.R
\name{bs3_navbar}
\alias{bs3_navbar}
\title{Custom navbar function for \code{\link[editR:editR-package]{editR}}}
\usage{
bs3_navbar(brand = "My website", navs)
}
\arguments{
\item{navs}{A list of UI elements to include in the navbar (e.g. \code{\link[shiny:actionLink]{actionLinks}}).}

\item{name}{The name of the navbar.}
}
\value{
This function return the html code for a Twitter Bootstrap 3 navbar.
}
\description{
This function creates custom Twitter Bootstrap 3 navbars for \code{\link[editR:editR-package]{editR}}.
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

