% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit_ui.R
\name{code_edit_ui}
\alias{code_edit_ui}
\title{Edit functions}
\usage{
code_edit_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified function in the appropriate function, template, or edition folder.
}
\description{
Module facilitating the quick creation or modification of functions, templates, or stylesheets used as or on documents.
}
\author{
Nicolas Mangin
}
