% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_server.R
\name{compare_server}
\alias{compare_server}
\title{Compare documents}
\usage{
compare_server(id, course_paths, courses)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{courses}{Reactive. Function containing a list of paths to other courses.}
}
\value{
Saved edited documents.
}
\description{
Module facilitating the comparison and edition of documents.
}
\author{
Nicolas Mangin
}
