% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ui.R
\name{compare_ui}
\alias{compare_ui}
\title{Compare documents}
\usage{
compare_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Saved edited documents.
}
\description{
Module facilitating the comparison and edition of documents.
}
\author{
Nicolas Mangin
}
