% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_curve.R
\name{display_curve}
\alias{display_curve}
\title{Display IRT curve}
\usage{
display_curve(selected_model)
}
\arguments{
\item{selected_model}{Tibble. Data produced by the function statistics_compute and stored in statistics$models for each question.}
}
\value{
Plot a curve.
}
\description{
Function drawing the IRT curve associated to a selected question.
}
\author{
Nicolas Mangin
}
