% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_question.R
\name{display_question}
\alias{display_question}
\title{Display the preview of a question.}
\usage{
display_question(selected, course_paths, test_parameters = NA)
}
\arguments{
\item{selected}{Tibble. Selected document.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{test_parameters}{Tibble. Test parameters if the document to display is a specific version of a question selected for a test.}
}
\value{
User interface showing the question within the interface, not as a modal.
}
\description{
Function creating a preview of a question embedded in a shiny dashboard, not as a model.
}
\author{
Nicolas Mangin
}
