% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation_server.R
\name{documents_translation_server}
\alias{documents_translation_server}
\title{Translate documents}
\usage{
documents_translation_server(id, preselection)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{preselection}{Tibble. List of pre-selected documents.}
}
\value{
Create and save documents' translations in a dedicated folder.
}
\description{
Module facilitating the translation of documents.
}
\author{
Nicolas Mangin
}
