% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation_ui.R
\name{documents_translation_ui}
\alias{documents_translation_ui}
\title{Translate documents}
\usage{
documents_translation_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Create and save documents' translations in a dedicated folder.
}
\description{
Module facilitating the translation of documents.
}
\author{
Nicolas Mangin
}
