% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_propositions_server.R
\name{edit_propositions_server}
\alias{edit_propositions_server}
\title{Edit propositions}
\usage{
edit_propositions_server(id, filtered, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{filtered}{Reactive. List of pre-selected documents.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Interface to edit propositions and their translations
}
\description{
Module facilitating the quick creation or modification of propositions.
}
\author{
Nicolas Mangin
}
