% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_propositions_ui.R
\name{edit_propositions_ui}
\alias{edit_propositions_ui}
\title{Edit documents}
\usage{
edit_propositions_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified document in the original documents folder.
}
\description{
Module facilitating the quick creation or modification of documents.
}
\author{
Nicolas Mangin
}
