% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_server.R
\name{edit_server}
\alias{edit_server}
\title{Edit documents}
\usage{
edit_server(id, filtered, course_data, intake, course_paths, doctype)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{filtered}{Reactive. List of pre-selected documents.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{intake}{Reactive. Function containing a list of documents as a classification tree compatible with jsTreeR}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{doctype}{Character. Whether the document is a "Note", "Page", "Slide", "Video", "Game", "Case" , or question}
}
\value{
Save the new or modified page in the folder "2_documents/main_language/".
}
\description{
Module facilitating the quick creation or modification of documents.
}
\author{
Nicolas Mangin
}
