\name{editR-package}
\alias{editR-package}
\docType{package}
\title{
A basic Rmarkdown editor with instant preview
}
\description{
editR is a basic Rmarkdown editor with instant previewing of your document. It 
allows you to create and edit \code{\link[<rmarkdown>:<rmarkdown>-package]{Rmarkdown}} 
documents while instantly previewing the result of your writing and coding. It 
also allows you to render your Rmarkdown file in any format permitted by the 
\code{\link[<rmarkdown>:<rmarkdown>-package]{Rmarkdown}} package. It was built 
using \code{\link[<shiny>:<shiny>-package]{shiny}} and its extensions 
\code{\link[<shinyBS>:<shinyBS>-package]{shinyBS}}, 
\code{\link[<shinyAce>:<shinyAce>-package]{shinyAce}},
\code{\link[<shinyFiles>:<shinyFiles>-package]{shinyFiles}}.
}
\details{
\tabular{ll}{
Package: \tab editR\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2015-01-01\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Simon Garnier <garnier@njit.edu> - \href{http://twitter.com/sjmgarnier}{@sjmgarnier}
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[<rmarkdown>:<rmarkdown>-package]{rmarkdown}}, \code{\link[<shiny>:<shiny>-package]{shiny}},
\code{\link[<shinyBS>:<shinyBS>-package]{shinyBS}}, \code{\link[<shinyAce>:<shinyAce>-package]{shinyAce}},
\code{\link[<shinyFiles>:<shinyFiles>-package]{shinyFiles}}
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
