% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get a document's tags}
\usage{
get_tags(path)
}
\arguments{
\item{path}{character. Path to the document for which meta-information should be retrieved.}
}
\value{
A single-row tibble with all the tags from a document.
}
\description{
Function reading a document and retrieving the tags in meta-information.
}
\author{
Nicolas Mangin
}
