% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_original_older.R
\name{is_original_older}
\alias{is_original_older}
\title{Check older file}
\usage{
is_original_older(original_file, alternative_file)
}
\arguments{
\item{original_file}{Character. Absolute path to the original file.}

\item{alternative_file}{Character. Absolute path to the alternative (comparison or translation) file.}
}
\value{
Logical. TRUE if the original file is older, FALSE otherwise.
}
\description{
Function returning whether the ortiginal file is older than the alternative file.
}
\author{
Nicolas Mangin
}
