% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_name.R
\name{make_new_name}
\alias{make_new_name}
\title{Create a document name}
\usage{
make_new_name(firstletter, course_paths)
}
\arguments{
\item{firstletter}{Character. "N" for notes, "P" for pages, "S" for slides, "V" for video scripts, "G" for games, "C" for cases, "Q" for questions.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Character. Name of a new document to be created.
}
\description{
Function listing the names of existing documents and making a new name by adding one to the highest existing increment. This ensures that no old number is assigned again.
}
\author{
Nicolas Mangin
}
