% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title_display.R
\name{make_title_display}
\alias{make_title_display}
\title{Create a title for a document}
\usage{
make_title_display(selected_document, course_data)
}
\arguments{
\item{selected_document}{Tibble. Document selected for display.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}
}
\value{
Character. Title to display for the document.
}
\description{
Function making a title with all relevant information for document displays
}
\author{
Nicolas Mangin
}
