% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree_path.R
\name{make_tree_path}
\alias{make_tree_path}
\title{Report the position of a document in a tree.}
\usage{
make_tree_path(doc, tree)
}
\arguments{
\item{doc}{Character. File name of the selected document.}

\item{tree}{Tibble. Tree in which the document should be positioned.}
}
\value{
Character string indicating the path to the document in the tree.
}
\description{
Function showing the path to the selected document in the selected tree.
}
\author{
Nicolas Mangin
}
