% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_yaml.R
\name{make_yaml}
\alias{make_yaml}
\title{Create yaml for various documents}
\usage{
make_yaml(selected, doctype)
}
\arguments{
\item{selected}{Tibble. Document for which the de YAML should be created.}

\item{doctype}{Character. Whether the document is a "Note", "Page", "Slide", "Script", "Game", or "Case" (Questions have no YAML).}
}
\value{
YAML as a character vector.
}
\description{
Make YAML to be inserted at the beginning of the document exported to be rendered via quarto.
}
\author{
Nicolas Mangin
}
