% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_new_item.R
\name{name_new_item}
\alias{name_new_item}
\title{New item name}
\usage{
name_new_item(existing_names = NULL, itmnbr = 1)
}
\arguments{
\item{existing_names}{Character vector. Item names already assigned.}

\item{itmnbr}{Integer. Number of items to create}
}
\value{
Character string. New item ID.
}
\description{
Function creating a new ID for a new item in the propositions database. Note that item IDs which have been removed will be reassigned.
}
\author{
Nicolas Mangin
}
