% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Prepare rmarkdown document for instant preview}
\usage{
preview(rmd)
}
\arguments{
\item{rmd}{The rmarkdown document to prepare for instant preview.}
}
\value{
This function return a rmarkdown document ready for instant preview in editR.
}
\description{
This function prepares a rmarkdown document edited with editR for instant preview.
It strips its front matter if it is present and add two extra line breaks at
the end to prevent editR to "misbehave" when a code chunk is created at the
very end of the document.
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

