% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_presentation.R
\name{publish_presentation}
\alias{publish_presentation}
\title{Export a slide presentation}
\usage{
publish_presentation(
  tree,
  selected_document,
  course_paths,
  translation = FALSE
)
}
\arguments{
\item{tree}{List.}

\item{selected_document}{Character. Name of the note destined to be published.}

\item{course_paths}{List.}

\item{translation}{Logical. Whether the file is a translation and should thus be published in a dedicated language subfolder.}
}
\value{
Write presentation in the folder "4_materials/presentation".
}
\description{
Function copying a presentation in the appropriate tree folder in 4_materials/presentation
}
\author{
Nicolas Mangin
}
