% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_server.R
\name{selection_server}
\alias{selection_server}
\title{Sleect from a vector}
\usage{
selection_server(id, selection_base)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{selection_base}{Reactive. Function containing a list of values from which to choose.}
}
\value{
A list of course data.
}
\description{
Module facilitating the selection of a specific value in a vector.
}
\author{
Nicolas Mangin
}
