% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_ui.R
\name{selection_ui}
\alias{selection_ui}
\title{Sleect from a vector}
\usage{
selection_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list of course data.
}
\description{
Module facilitating the selection of a specific value in a vector.
}
\author{
Nicolas Mangin
}
