% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate a character string}
\usage{
translate(progress = NA, text, langiso)
}
\arguments{
\item{progress}{Charater. Identify the item or row which is being translated.}

\item{text}{Charater. Character string which is being translated.}

\item{langiso}{Character. ISO2 code of the language in which propositions and explanations should be translated.}
}
\value{
Tibble. Translated propositions and explanations.
}
\description{
Function translating character strings.
}
\author{
Nicolas Mangin
}
