% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_document.R
\name{translate_document}
\alias{translate_document}
\title{Translate documents}
\usage{
translate_document(filepath, langiso)
}
\arguments{
\item{filepath}{Character. Path to the original file to be translated.}

\item{langiso}{Character. ISO2 code of the language in lower cases.}
}
\value{
Tibble. Translated propositions and explanations.
}
\description{
Function translating non-translated propositions and explanations in the selected language.
}
\author{
Nicolas Mangin
}
