% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_propositions.R
\name{translate_propositions}
\alias{translate_propositions}
\title{Translate propositions}
\usage{
translate_propositions(propositions, translations, langiso)
}
\arguments{
\item{propositions}{Tibble. All the propositions.}

\item{translations}{Tibble. Table associating translations to propositions.}

\item{langiso}{Character. ISO2 code of the language in which propositions and explanations should be translated.}
}
\value{
Tibble. Translated propositions and explanations.
}
\description{
Function translating non-translated propositions and explanations in the selected language.
}
\author{
Nicolas Mangin
}
