% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_server.R
\name{translate_server}
\alias{translate_server}
\title{Translate documents}
\usage{
translate_server(id, filtered, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{filtered}{Reactive. List of pre-selected documents.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Create and save documents' translations in a dedicated folder.
}
\description{
Module facilitating the translation of documents.
}
\author{
Nicolas Mangin
}
