% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_ui.R
\name{translate_ui}
\alias{translate_ui}
\title{Translate documents}
\usage{
translate_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Create and save documents' translations in the dedicated basis subfolder.
}
\description{
Module facilitating the translation of documents.
}
\author{
Nicolas Mangin
}
