% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_form.R
\name{add_form}
\alias{add_form}
\title{Add a form for comments}
\usage{
add_form(
  sourcename,
  filename,
  message =
    "Please let us know below if you have questions or comments about this page:  "
)
}
\arguments{
\item{sourcename}{Character.}

\item{filename}{Character.}

\item{message}{Character.}
}
\value{
Write the HTML code for the form in the page
}
\description{
Function writing the HTML code to give the possibility to make comments.
}
\author{
Nicolas Mangin
}
