% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rating.R
\name{add_rating}
\alias{add_rating}
\title{Add star rating}
\usage{
add_rating(
  sourcename,
  filename,
  message =
    "Please let us know below how clear and understandable this page was for you:  "
)
}
\arguments{
\item{sourcename}{Character.}

\item{filename}{Character.}

\item{message}{Character.}
}
\value{
Write the HTML code for the star rating in the page
}
\description{
Function writing the HTML code to include the page star rating.
}
\author{
Nicolas Mangin
}
