% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_slide.R
\name{add_slide}
\alias{add_slide}
\title{Insert a slide}
\usage{
add_slide(
  text = "",
  format = "titled",
  colorback = "#FFFFFF",
  image = "",
  video = "",
  transback = "slide",
  transdata = "fade",
  state = "dimbg100"
)
}
\arguments{
\item{text}{Character. Text of the slide.}

\item{format}{Character. "section" for a section, "titled" for a slide with title, and "untitled" for a slide without title.}

\item{colorback}{Character. Background color.}

\item{image}{Character. Address of the background image.}

\item{video}{Character. Address of the background video.}

\item{transback}{Character. Type of transition for the background: fade, slide, convex, concave, or zoom.}

\item{transdata}{Character. Type of transition for data: fade, slide, convex, concave, or zoom.}

\item{state}{Character. "dimbg140" to "dimbg20" (20 by 20) for a decreasing brightness and "blurbg5" for a blur effect.}
}
\value{
Character. Write the first two rows of a RevealJS slide formatted for rmarkdown.
}
\description{
Function writing a revealjs slide according to the user specifications.
}
\author{
Nicolas Mangin
}
