% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_video.R
\name{add_video}
\alias{add_video}
\title{Embed video}
\usage{
add_video(vid, width = "100\%", height = "100\%", start = 0, host = "Youtube")
}
\arguments{
\item{vid}{Character. ID of the video.}

\item{width}{Character. Width of the video.}

\item{height}{Character. Height of the video.}

\item{start}{Numeric. Time at which the video should start.}

\item{host}{Character. Name of the hosting platform. For now, only "Youtube" is supported.}
}
\value{
Write the HTML code embedding the videos.
}
\description{
Function embedding a video as an ifram in an html page
}
\author{
Nicolas Mangin
}
