% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editR.R
\name{editR}
\alias{editR}
\title{Edit Rmarkdown files with instant preview}
\usage{
editR(file = NULL, port = NULL)
}
\arguments{
\item{file}{The path (as a character string) to an existing .Rmd/.md file, or 
to a new .Rmd/.md file. If the file does not exist, it will be created when the 
GUI starts.]}
}
\value{
This function does not return anything.
}
\description{
This function creates a GUI to edit \code{\link[rmarkdown:rmarkdown-package]{Rmarkdown}} 
documents and display an instant preview of the document as it is being 
written. Rmarkdown documents can be rendered directly from the GUI.
}
\details{
If called from RStudio on Mac and Linux, it will open in the internal 
RStudio browser. If called from a terminal or on Windows, it will open in your 
default internet browser. editR should work without any problem with recent 
versions of most internet browsers. It is likely to break with older versions.
}
\examples{
# Create and edit a new Rmarkdown document in the current working directory
editR("example.Rmd")

}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

