% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_script.R
\name{publish_script}
\alias{publish_script}
\title{Format a video as a script}
\usage{
publish_script(selected_document, course_paths, translation = FALSE)
}
\arguments{
\item{selected_document}{Character. Name of the note destined to be published.}

\item{course_paths}{List.}

\item{translation}{Logical. Whether the source document is in the folder "translated" rather than "original".}
}
\value{
Write post in the folder "4_materials/blog".
}
\description{
Function taking only the lines supposed to be uttered and exporting them into a .txt file which can be read by a prompter.
}
\author{
Nicolas Mangin
}
