% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_textbook.R
\name{publish_textbook}
\alias{publish_textbook}
\title{Compile a textbook}
\usage{
publish_textbook(tree, course_paths, languages)
}
\arguments{
\item{tree}{List. Selected tree to retrieve the position of the presentation.}

\item{course_paths}{List. List of paths to either databases or documents.}

\item{languages}{Tibble. List of managed languages.}
}
\value{
Write folders and pages to disk in the textbook folder.
}
\description{
Function loading a tree and compiling the corresponding textbook. Pages are formatted and organized so that they can then be added to a quarto website.
}
\author{
Nicolas Mangin
}
