% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_textbook.R
\name{publish_textbook}
\alias{publish_textbook}
\title{Compile a textbook}
\usage{
publish_textbook(tree, course_paths, languages)
}
\arguments{
\item{tree}{List.}

\item{course_paths}{List.}

\item{languages}{Tibble}
}
\value{
Write folders and pages to disk in the folder "4_materials/textbooks".
}
\description{
Function loading a tree and compiling the corresponding textbook. Pages are formatted and organized so that they can then be added to a quarto website.
}
\author{
Nicolas Mangin
}
