% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_exercise.R
\name{translate_exercise}
\alias{translate_exercise}
\title{Translate exercise}
\usage{
translate_exercise(exercise, translations, languiso)
}
\arguments{
\item{exercise}{Tibble. Exercise table in the main language.}

\item{translations}{Tibble. Table associating translations to propositions.}

\item{languiso}{Character. ISO2 code of the language.}
}
\value{
Tibble. Tanslated exercise.
}
\description{
Function replacing propositions and explanation by their translations
}
\author{
Nicolas Mangin
}
