% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_document.R
\name{view_document}
\alias{view_document}
\title{Preview a document.}
\usage{
view_document(selected, original, course_paths)
}
\arguments{
\item{selected}{Tibble. Selected document.}

\item{original}{Logical. Whether the document is the original (TRUE) or a translation (FALSE).}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
User interface showing the formatted document.
}
\description{
Function creating the user interface to see formatted documents.
}
\author{
Nicolas Mangin
}
